+++ xroad-catalog-persistence/src/main/java/fi/vrk/xroad/catalog/persistence/repository/MemberRepository.java
@@ -69,4 +69,7 @@ public interface MemberRepository extends CrudRepository<Member, Long> {
                                   @Param("memberClass") String memberClass,
                                   @Param("memberCode") String memberCode);
 
+    @Query(value = "SELECT 1", nativeQuery = true)
+    Integer checkConnection();
+
 }